removed the now obsolete newsAlerts apis
